+++ Changelog.md
@@ -1,6 +1,20 @@
-[]Changelog[]
+----------Changelog----------
+[v0.3.0] (11/18/23)
+- Reorganized the user interface.
+- If a transformation button does not have an icon specified, it will show text for the transformation instead. The transformation icons folder is now optional.
+- Transformation buttons now have a background image that is always displayed behind their icon/text.
+- You can now overwrite various images, including:
+    - The missing image placeholder image.
+    - The transformation button background image.
+    - The undo button image.
+    - The exit button image.
+- The default undo and exit buttons and missing image placeholder have been redesigned.
+- Add the Sphere Demo database, a more bare-bones example of a functioning database.
+- Updated the Draden Demo database to V3, adding custom button and placeholder images.
+- Updated database setup instructions.
+- Updated the README.
 
-[v0.2.0] ()
+[v0.2.0] (11/11/23)
 - Added small green dots that appear near a transformation button when clicking that button will result in an image instead of a placeholder.
 - Added a stack of text showing which transformations have been applied, with the most recent on top.
 - Added text that displays during gameplay showing the character's name and database author's name as specified in the "Names.txt" file.